Effect class separation
Separated the various Effect classes into their own files
